requirements.txt
pypdf
llama-cpp-python
sentence-transformers
fastapi
uvicorn
websockets
python-multipart
cryptography
chromadb